requirements.txt
langchain_groq
langserve
fastapi
langchain_core
python-dotenv
uvicorn
sse_starlette